move test from integration to functional
